fixed homepage checkbox category filter
